@@ -2,7 +2,8 @@
   "name": "union-frontend",
   "version": "0.0.0",
   "scripts": {
-    "dev": "npm run copy_deps && vite",
+    "dev": "npm run copy_deps && npm run dev-frontend",
+    "dev-frontend": "vite",
     "build": "npm run copy_deps && npm run build-frontend",
     "build-frontend": "vite build",
     "copy_deps": "./build_scripts/pre_copy_deps.sh",
@@ -14,7 +15,7 @@
   },
   "dependencies": {
     "@union/serialize": "*",
-    "@union/components": "*",
+    "@union/candid-parser": "*",
     "@dfinity/agent": "^0.10.4",
     "@dfinity/auth-client": "^0.10.4",
     "@dfinity/authentication": "^0.10.4",
@@ -26,21 +27,21 @@
     "history": "5.3.0",
     "moment": "^2.29.1",
     "qs": "^6.9.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.9",
+    "react-dom": "^18.0.9",
     "react-hook-form": "^7.29.0",
     "react-router-dom": "6.2.2",
     "react-transition-group": "^4.4.2",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.5"
   },
   "devDependencies": {
     "@types/moment": "^2.13.0",
     "@types/node": "^17.0.21",
     "@types/qs": "^6.9.4",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
     "@types/react-transition-group": "^4.4.4",
-    "@types/styled-components": "5.1.18",
+    "@types/styled-components": "5.1.25",
     "@typescript-eslint/eslint-plugin": "~4.4.1",
     "@typescript-eslint/parser": "~4.4.1",
     "@vitejs/plugin-react-refresh": "^1.3.2",
@@ -54,7 +55,7 @@
     "eslint-plugin-react-hooks": "^4.1.2",
     "prettier": "~2.1.2",
     "rollup-plugin-copy": "^3.4.0",
-    "typescript": "^4.2.4",
+    "typescript": "4.6.4",
     "vite": "^2.6.14",
     "vite-plugin-checker": "^0.4.5"
   }
