@@ -1,10 +1,7 @@
-import React, { useEffect, useMemo } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import { NavLink, useParams } from 'react-router-dom';
+import { NavLink } from 'react-router-dom';
 import { Text } from '@union/components';
-import { useDeployer, useUnion } from 'services';
-import { checkPrincipal } from 'toolkit';
-import { Principal } from '@dfinity/principal';
 import { LoginButton } from '../Auth/LoginButton';
 
 const Item = styled(Text)`
@@ -12,8 +9,8 @@ const Item = styled(Text)`
   color: #575757;
 
   &.active {
-    color: black;
-    font-weight: 500;
+    color: ${({ theme }) => theme.colors.dark};
+    font-weight: 600;
   }
 `;
 
@@ -36,34 +33,12 @@ const Container = styled.div`
   justify-content: space-between;
   align-items: center;
   padding: 8px 16px;
-  border-bottom: 1px solid ${({ theme }) => theme.colors.dark[900]};
+  border-bottom: 1px solid ${({ theme }) => theme.colors.dark};
 `;
 
 export interface HeaderProps extends IClassName {}
 
 export function Header(p: HeaderProps) {
-  const param = useParams();
-  const location = param['*'] || '';
-
-  const isInsideWallet = location.startsWith('wallet/');
-  const walletId = location.split('wallet/')[1]?.split('/')[0] || '';
-  const { data, canister } = useUnion(walletId ? Principal.from(walletId) : Principal.anonymous());
-
-  useEffect(() => {
-    if (!checkPrincipal(walletId)) {
-      return;
-    }
-    canister.get_my_profile();
-  }, [walletId, canister]);
-
-  const profileName = useMemo(() => {
-    if (!isInsideWallet || !walletId) {
-      return '';
-    }
-    // TODO get groups
-    return data.get_my_profile?.profile.name || '';
-  }, [data.get_my_profile, isInsideWallet && walletId]);
-
   return (
     <Container {...p}>
       <Items>
@@ -80,7 +55,7 @@ export function Header(p: HeaderProps) {
           Explore
         </Item> */}
       </Items>
-      <LoginButton name={profileName} />
+      <LoginButton />
     </Container>
   );
 }
