@@ -1,14 +1,12 @@
-import React, { useCallback, useEffect, useMemo } from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
-import { PageWrapper, Text, SubmitButton as SB, ImageFile as IF } from '@union/components';
-import { useDeployer, useUnion } from 'services';
+import { PageWrapper, Text, SubmitButton as SB, Field as F } from '@union/components';
+import { useUnion } from 'services';
 import { NavLink } from 'react-router-dom';
-import { Principal } from '@dfinity/principal';
 import moment from 'moment';
 import { useCurrentUnion } from '../context';
 
-const ImageFile = styled(IF)``;
-const Field = styled(Text)``;
+const Field = styled(F)``;
 const Button = styled(SB)``;
 const Controls = styled.div`
   display: flex;
@@ -20,16 +18,17 @@ const Controls = styled.div`
   }
 `;
 const Container = styled(PageWrapper)`
-  ${ImageFile} {
-    height: 100px;
-    width: 100px;
-  }
   ${Button} {
     align-self: flex-start;
   }
 
-  ${Field}, ${ImageFile} {
+  & > ${Field} {
     margin-bottom: 8px;
+  }
+
+  ${Field} ${Field} {
+    padding-left: 8px;
+    border-left: 1px solid ${({ theme }) => theme.colors.grey};
   }
 `;
 
@@ -41,11 +40,9 @@ export interface InfoProps {
 export const Info = ({ ...p }: InfoProps) => {
   const { principal } = useCurrentUnion();
   const { canister, fetching, data } = useUnion(principal);
-  const deployer = useDeployer(process.env.UNION_DEPLOYER_CANISTER_ID);
 
   useEffect(() => {
     canister.get_settings();
-    deployer.canister.get_instances({ ids: [Principal.from(principal)] });
   }, []);
 
   const settings = data.get_settings?.settings || null;
@@ -63,30 +60,43 @@ export const Info = ({ ...p }: InfoProps) => {
       {!!fetching.get_settings && <Text>fetching...</Text>}
       {settings && (
         <>
-          <Field>ID: {principal.toString()}</Field>
-          <Field>Name: {settings.name}</Field>
-          <Field>Description: {settings.description}</Field>
-          <Field>Version: {deployer.data.get_instances?.instances[0].binary_version || '?'}</Field>
+          <Field title='ID' align='row'>
+            {principal.toString()}
+          </Field>
+          <Field title='Name' align='row'>
+            {settings.name}
+          </Field>
+          <Field title='Description' align='row'>
+            {settings.description}
+          </Field>
+          <Field title='Version' align='row'>
+            ?
+          </Field>
+          <Field title='Balance' align='row'>
+            ?
+          </Field>
+          <Field title='Storage' align='row'>
+            ?
+          </Field>
           {settings.history_ledgers.map((ledger, i) => {
             const timestamp = moment(Number(ledger.timestamp) / 10 ** 6).format(
               'DD-MM-YY HH:mm:SS',
             );
 
             return (
-              <Field key={String(i)}>
-                <Field>Ledger #{i}</Field>
+              <Field key={String(i)} title={`Ledger #${i}`}>
                 {ledger.records.map((r, j) => (
-                  <Field key={r.toString()}>
-                    Record #{j}: {r.toString()}
+                  <Field key={r.toString()} title={`Record #${j}`} align='row'>
+                    {r.toString()}
                   </Field>
                 ))}
 
-                <Field>Timestamp: {timestamp}</Field>
+                <Field title='Timestamp' align='row'>
+                  Timestamp: {timestamp}
+                </Field>
               </Field>
             );
           })}
-          {/* <Field>Balance: ?</Field>
-          <Field>Storage: ?</Field> */}
         </>
       )}
     </Container>
