@@ -1,9 +1,11 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect } from 'react';
 import styled from 'styled-components';
-import { PageWrapper, SubmitButton as B } from '@union/components';
+import { PageWrapper, SubmitButton as B, Text } from '@union/components';
 import { useGateway, useUnion } from 'services';
-import { HAS_PROFILE_GROUP_ID } from '../../../envs';
+import { useNavigate } from 'react-router-dom';
+import { HAS_PROFILE_GROUP_ID } from 'envs';
 import { useCurrentUnion } from '../context';
+import { Groups } from './Groups';
 
 const Button = styled(B)``;
 
@@ -18,7 +20,7 @@ const Controls = styled.div`
 `;
 
 const Container = styled(PageWrapper)`
-  ${Controls} {
+  ${Controls}, ${Groups} {
     margin-bottom: 24px;
   }
 `;
@@ -29,18 +31,20 @@ export interface ProfileProps {
 }
 
 export const Profile = ({ ...p }: ProfileProps) => {
-  const { principal, profile, groups } = useCurrentUnion();
+  const nav = useNavigate();
+  const { principal, profile, groups, fetchMyData } = useCurrentUnion();
   const { canister, data } = useUnion(principal);
   const gateway = useGateway(process.env.GATEWAY_CANISTER_ID);
 
   useEffect(() => {
+    fetchMyData();
     canister.get_my_unaccepted_group_shares_balance({
       group_id: HAS_PROFILE_GROUP_ID,
     });
   }, []);
 
   const handleAccept = useCallback(async () => {
-    const qty = data.get_my_unaccepted_group_shares_balance?.balance || 0n;
+    const qty = data.get_my_unaccepted_group_shares_balance?.balance || BigInt(0);
 
     if (!qty) {
       return;
@@ -49,6 +53,9 @@ export const Profile = ({ ...p }: ProfileProps) => {
     await canister.accept_my_group_shares({ group_id: HAS_PROFILE_GROUP_ID, qty });
 
     await gateway.canister.attach_to_union_wallet({ union_wallet_id: principal });
+    await canister.get_my_unaccepted_group_shares_balance({
+      group_id: HAS_PROFILE_GROUP_ID,
+    });
   }, [gateway, groups, principal, data.get_my_unaccepted_group_shares_balance?.balance]);
 
   return (
@@ -57,11 +64,11 @@ export const Profile = ({ ...p }: ProfileProps) => {
         {!!data.get_my_unaccepted_group_shares_balance?.balance && (
           <Button onClick={handleAccept}>Accept invite</Button>
         )}
+        <Button onClick={() => nav('change')}>Change profile</Button>
       </Controls>
-      {/* {!!fetching.list_batches && <Text>fetching</Text>} */}
-      {/* {!fetching.list_batches && !batches.length && <Text>Batches does not exist</Text>} */}
-      {profile?.name}
-      {profile?.description}
+      <Text>{profile?.name}</Text>
+      <Text>{profile?.description}</Text>
+      <Groups />
     </Container>
   );
 };
