@@ -1,4 +1,4 @@
 export * from './gateway';
 export * from './deployer';
-export * from './wallet';
+export * from './union';
 export * from './management';
